revert

+++ vkquake/build-vkquake.sh
@@ -37,11 +37,11 @@ fi
 
 if [[ "$arg1" == "--testing" ]]; then
 
-	REPO_FOLDER="/home/mikeyd/packaging/static-builds/testing"
+	REPO_FOLDER="/home/mikeyd/packaging/steamos-tools/incoming_testing"
 
 else
 
-	REPO_FOLDER="/home/mikeyd/packaging/static-builds/stable"
+	REPO_FOLDER="/home/mikeyd/packaging/steamos-tools/incoming"
 
 fi
 # upstream vars
@@ -138,51 +138,100 @@ main()
 	# enter source dir
 	cd "${GIT_DIR}"
 
+	echo -e "\n==> Updating changelog"
+	sleep 2s
+
+	# update changelog with dch
+	# "Update to the latest commit ${latest_commit}"
+	if [[ -f "debian/changelog" ]]; then
+
+		dch -p --force-distribution -v "${epoch}:${PKGVER}+${PKGSUFFIX}-${PKGREV}" \
+		--package "${PKGNAME}" -D "${DIST}" -u "${URGENCY}" \
+		"Update to the latest commit ${latest_commit}"
+		nano "debian/changelog"
+
+	else
+
+		dch -p --create --force-distribution -v "${epoch}:${PKGVER}+${PKGSUFFIX}-${PKGREV}" \
+		--package "${PKGNAME}" -D "${DIST}" -u "${URGENCY}" "Initial build"
+		nano "debian/changelog"
+
+	fi
 
 	#################################################
 	# Build Debian package
 	#################################################
 
-	echo -e "\n==> Building static Debian package ${PKGNAME} from source\n"
+	echo -e "\n==> Building Debian package ${PKGNAME} from source\n"
 	sleep 2s
 
-	cd vkQuake/Quake
-	make || echo -e "\n==EROR==\nMake failed!"
+	USENETWORK=$NETWORK DIST=$DIST ARCH=$ARCH ${BUILDER} ${BUILDOPTS}
 
 	#################################################
 	# Cleanup
 	#################################################
 
+	# clean up dirs
 
-	echo -e "\n==> Would you like to transfer the archive? [y/n]"
-	sleep 0.5s
-	# capture command
-	read -erp "Choice: " transfer_choice
+	# note time ended
+	time_end=$(date +%s)
+	time_stamp_end=(`date +"%T"`)
+	runtime=$(echo "scale=2; ($time_end-$TIME_START) / 60 " | bc)
 
-	if [[ "$transfer_choice" == "y" ]]; then
+	# output finish
+	echo -e "\nTime started: ${TIME_STAMP_START}"
+	echo -e "Time started: ${time_stamp_end}"
+	echo -e "Total Runtime (minutes): $runtime\n"
 
-		# copy files to remote server
-		rsync -arv --info=progress2 -e "ssh -p ${REMOTE_PORT}" \
-		--filter="merge ${HOME}/.config/SteamOS-Tools/repo-filter.txt" \
-		${BUILD_TMP}/ ${REMOTE_USER}@${REMOTE_HOST}:${REPO_FOLDER}
 
-		# uplaod local repo changelog
-		cp "${GIT_DIR}/debian/changelog" "${SCRIPTDIR}/debian"
+	# assign value to build folder for exit warning below
+	build_folder=$(ls -l | grep "^d" | cut -d ' ' -f12)
 
-	elif [[ "$transfer_choice" == "n" ]]; then
-		echo -e "Upload not requested\n"
+	# inform user of packages
+	cat<<- EOF
+	#################################################################
+	If package was built without errors you will see it below.
+	If you don't, please check build dependency errors listed above.
+	#################################################################
+
+	EOF
+
+	echo -e "Showing contents of: ${BUILD_TMP}: \n"
+	ls "${BUILD_TMP}" | grep -E *${PKGVER}*
+
+	# Ask to transfer files if debian binries are built
+	# Exit out with log link to reivew if things fail.
+
+	if [[ $(ls "${BUILD_TMP}" | grep -w "deb" | wc -l) -gt 0 ]]; then
+
+		echo -e "\n==> Would you like to transfer any packages that were built? [y/n]"
+		sleep 0.5s
+		# capture command
+		read -erp "Choice: " transfer_choice
+
+		if [[ "$transfer_choice" == "y" ]]; then
+
+			# copy files to remote server
+			rsync -arv --info=progress2 -e "ssh -p ${REMOTE_PORT}" \
+			--filter="merge ${HOME}/.config/SteamOS-Tools/repo-filter.txt" \
+			${BUILD_TMP}/ ${REMOTE_USER}@${REMOTE_HOST}:${REPO_FOLDER}
+
+			# uplaod local repo changelog
+			cp "${GIT_DIR}/debian/changelog" "${SCRIPTDIR}/debian"
+
+		elif [[ "$transfer_choice" == "n" ]]; then
+			echo -e "Upload not requested\n"
+		fi
+
+	else
+
+		# Output log file to sprunge (pastebin) for review
+		echo -e "\n==OH NO!==\nIt appears the build has failed. See below log file:"
+		cat ${BUILD_TMP}/${PKGNAME}*.build | curl -F 'sprunge=<-' http://sprunge.us
+
 	fi
-
-else
-
-	# Output log file to sprunge (pastebin) for review
-	echo -e "\n==OH NO!==\nIt appears the build has failed. See below log file:"
-	cat ${BUILD_TMP}/${PKGNAME}*.build | curl -F 'sprunge=<-' http://sprunge.us
-
-fi
 
 }
 
 # start main
 main
-
